Android 16.0.0 release 4

* tag 'android-16.0.0_r4': (6818 commits)
  Fix skip epoll flag behavior
  Fix predictive back system animation stuck
  notifyFinished: Send requestedVisibleTypes before update animatingTypes
  Import translations. DO NOT MERGE ANYWHERE
  Import translations. DO NOT MERGE ANYWHERE
  Import translations. DO NOT MERGE ANYWHERE
  Import translations. DO NOT MERGE ANYWHERE
  Import translations. DO NOT MERGE ANYWHERE
  Import translations. DO NOT MERGE ANYWHERE
  Import translations. DO NOT MERGE ANYWHERE
  Import translations. DO NOT MERGE ANYWHERE
  Import translations. DO NOT MERGE ANYWHERE
  Import translations. DO NOT MERGE ANYWHERE
  Import translations. DO NOT MERGE ANYWHERE
  Do not select displays that cannot host the activities as fallback
  Add a check for enable_display_content_mode_management flag when checking display.canHostTasks()
  Don't allow camera gesture while SIM locked
  Remove bright red opaque status bar.
  Reorder EXECUTE_APP_FUNCTIONS permission declaration
  Validate displayName for AssociationRequest
  ...

+++ libs/usb/tests/accessorytest/hid.c
@@ -47,7 +47,7 @@ static void milli_sleep(int millis) {
 }
 
 static void* hid_thread(void* arg) {
-    int fd = (int)arg;
+    int fd = (int)(uintptr_t)arg;
     char buffer[4096];
     int id, ret, offset;
     struct usb_device *device;
